+++ lib/src/xml/nodes/document_fragment.dart
@@ -2,10 +2,10 @@ import 'package:petitparser/petitparser.dart';
 
 import '../entities/default_mapping.dart';
 import '../entities/entity_mapping.dart';
+import '../exceptions/parser_exception.dart';
 import '../mixins/has_children.dart';
 import '../parser.dart';
 import '../utils/cache.dart';
-import '../utils/exceptions.dart';
 import '../utils/node_type.dart';
 import '../visitors/visitor.dart';
 import 'node.dart';
@@ -23,13 +23,8 @@ class XmlDocumentFragment extends XmlNode with XmlHasChildren<XmlNode> {
     final parser = documentFragmentParserCache[mapping];
     final result = parser.parse(input);
     if (result.isFailure) {
-      final lineAndColumn =
-          Token.lineAndColumnOf(result.buffer, result.position);
       throw XmlParserException(result.message,
-          buffer: result.buffer,
-          position: result.position,
-          line: lineAndColumn[0],
-          column: lineAndColumn[1]);
+          buffer: result.buffer, position: result.position);
     }
     return result.value;
   }