+++ lib/src/xml/utils/node_list.dart
@@ -1,8 +1,9 @@
 import 'package:collection/collection.dart' show DelegatingList;
 import 'package:meta/meta.dart';
 
+import '../exceptions/parent_exception.dart';
+import '../exceptions/type_exception.dart';
 import '../nodes/node.dart';
-import 'exceptions.dart';
 import 'node_type.dart';
 
 /// Mutable list of XmlNodes, manages the parenting of the nodes.