+++ lib/src/xml_events/event.dart
@@ -1,10 +1,13 @@
 import '../xml/utils/node_type.dart';
+import 'annotations/has_buffer.dart';
+import 'annotations/has_location.dart';
+import 'annotations/has_parent.dart';
 import 'converters/event_encoder.dart';
-import 'utils/parented.dart';
 import 'visitor.dart';
 
 /// Immutable base class for all events.
-abstract class XmlEvent with XmlParented {
+abstract class XmlEvent with XmlHasParent, XmlHasLocation, XmlHasBuffer {
+  /// Default constructor for an event.
   XmlEvent();
 
   /// Return the node type of this node.