+++ lib/src/xml_events/iterable.dart
@@ -3,11 +3,29 @@ import 'event.dart';
 import 'iterator.dart';
 
 class XmlEventIterable extends Iterable<XmlEvent> {
-  XmlEventIterable(this.input, this.entityMapping);
+  XmlEventIterable(
+    this.input, {
+    required this.entityMapping,
+    required this.withBuffer,
+    required this.withLocation,
+    required this.withParent,
+    required this.validateNesting,
+  });
 
   final String input;
   final XmlEntityMapping entityMapping;
+  final bool validateNesting;
+  final bool withBuffer;
+  final bool withLocation;
+  final bool withParent;
 
   @override
-  Iterator<XmlEvent> get iterator => XmlEventIterator(input, entityMapping);
+  Iterator<XmlEvent> get iterator => XmlEventIterator(
+        input,
+        entityMapping: entityMapping,
+        validateNesting: validateNesting,
+        withBuffer: withBuffer,
+        withLocation: withLocation,
+        withParent: withParent,
+      );
 }