+++ lib/src/xml_events/iterator.dart
@@ -1,22 +1,36 @@
-import 'package:petitparser/petitparser.dart'
-    show Parser, Result, Success, Token;
+import 'package:petitparser/petitparser.dart' show Parser, Result, Failure;
 
 import '../xml/entities/entity_mapping.dart';
-import '../xml/utils/exceptions.dart';
+import '../xml/exceptions/parser_exception.dart';
+import 'annotations/annotator.dart';
 import 'event.dart';
 import 'parser.dart';
 
 class XmlEventIterator extends Iterator<XmlEvent> {
-  XmlEventIterator(String input, XmlEntityMapping entityMapping)
-      : _eventParser = eventParserCache[entityMapping],
-        _context = Success(input, 0, null);
+  XmlEventIterator(
+    String input, {
+    required XmlEntityMapping entityMapping,
+    required bool withBuffer,
+    required bool withLocation,
+    required bool withParent,
+    required bool validateNesting,
+  })  : _eventParser = eventParserCache[entityMapping],
+        _annotator = Annotator(
+          validateNesting: validateNesting,
+          withBuffer: withBuffer,
+          withLocation: withLocation,
+          withParent: withParent,
+        ),
+        _context = Failure<XmlEvent>(input, 0, '');
 
-  final Parser _eventParser;
-  Result? _context;
-  late XmlEvent _current;
+  final Parser<XmlEvent> _eventParser;
+  final Annotator _annotator;
+
+  Result<XmlEvent>? _context;
+  XmlEvent? _current;
 
   @override
-  XmlEvent get current => _current;
+  XmlEvent get current => _current!;
 
   @override
   bool moveNext() {
@@ -26,20 +40,17 @@ class XmlEventIterator extends Iterator<XmlEvent> {
       if (result.isSuccess) {
         _context = result;
         _current = result.value;
+        _annotator.annotate(context, result, result.value);
         return true;
       } else if (context.position < context.buffer.length) {
         // In case of an error, skip one character and throw an exception.
         _context = context.failure(result.message, context.position + 1);
-        final lineAndColumn =
-            Token.lineAndColumnOf(result.buffer, result.position);
         throw XmlParserException(result.message,
-            buffer: result.buffer,
-            position: result.position,
-            line: lineAndColumn[0],
-            column: lineAndColumn[1]);
+            buffer: result.buffer, position: result.position);
       } else {
         // In case of reaching the end, terminate the iterator.
         _context = null;
+        _annotator.close(context);
         return false;
       }
     }