+++ lib/src/xml_events/streams/normalizer.dart
@@ -38,7 +38,7 @@ class _XmlNormalizeEventsSink extends ChunkedConversionSink<List<XmlEvent>> {
       final event1 = buffer[i], event2 = buffer[i + 1];
       if (event1 is XmlTextEvent && event2 is XmlTextEvent) {
         final event = XmlTextEvent(event1.text + event2.text);
-        event.attachParentEvent(event1.parentEvent);
+        // event.attachParentEvent(event1.parentEvent);
         buffer[i] = event;
         buffer.removeAt(i + 1);
       } else {