+++ lib/src/xml_events/streams/subtree_selector.dart
@@ -1,6 +1,6 @@
 import 'dart:convert' show ChunkedConversionSink;
 
-import '../../xml/utils/exceptions.dart';
+import '../../xml/exceptions/tag_exception.dart';
 import '../../xml/utils/functions.dart';
 import '../event.dart';
 import '../events/end_element.dart';